+++ examples/surreal_ql/define_field.py
@@ -13,6 +13,8 @@ with Database("http://127.0.0.1:8000", 'test', 'test', credentials=('user_db', '
     print(db.define_field("field", "user").type("bool", is_flexible=True))
     # DEFINE FIELD locked ON TABLE user TYPE bool DEFAULT false;
     print(db.define_field("locked", "user").type("bool").default("false"))
+    # DEFINE FIELD locked ON TABLE user TYPE bool DEFAULT ALWAYS false;
+    print(db.define_field("locked", "user").type("bool").default("false", always=True))
     # DEFINE FIELD locked ON TABLE user TYPE bool DEFAULT false COMMENT "Some comment";
     print(db.define_field("locked", "user").type("bool").default("false").comment("Some comment"))
     # DEFINE FIELD updated ON TABLE resource DEFAULT time::now();
@@ -25,3 +27,18 @@ with Database("http://127.0.0.1:8000", 'test', 'test', credentials=('user_db', '
     print(db.define_field("comment", "resource").type("string", is_flexible=True).permissions_full())
     # DEFINE FIELD comment ON TABLE resource FLEXIBLE TYPE string PERMISSIONS FULL COMMENT "text";
     print(db.define_field("comment", "resource").type("string", is_flexible=True).permissions_full().comment("text"))
+    #  DEFINE FIELD comics ON person TYPE option<array<record<comic_book>>> REFERENCE;
+    print(db.define_field("comics", "person"). type("option<array<record<comic_book>>>").reference())
+    # DEFINE FIELD comics ON TABLE person TYPE option<array<record<comic_book>>> REFERENCE ON DELETE IGNORE;
+    print(db.define_field("comics", "person"). type("option<array<record<comic_book>>>").reference_ignore())
+    update = """{
+    UPDATE $this SET
+        deleted_comments += $reference,
+        comments -= $reference;
+}"""
+#     DEFINE FIELD comics ON TABLE person TYPE option<array<record<comic_book>>> REFERENCE ON DELETE THEN {
+#     UPDATE $this SET
+#         deleted_comments += $reference,
+#         comments -= $reference;
+# };
+    print(db.define_field("comics", "person").type("option<array<record<comic_book>>>").reference_then(update))