+++ src/surrealist/ql/statements/__init__.py
@@ -1,6 +1,8 @@
 from .create import Create
-from .define import DefineEvent, DefineParam
+from .define import DefineParam, DefineEvent, DefineIndex, DefineScope, DefineTable, DefineToken
+from .define_access import DefineAccessBearer, DefineAccessJwt, DefineAccessRecord
 from .define_analyzer import DefineAnalyzer
+from .define_field import DefineField
 from .define_user import DefineUser
 from .delete import Delete
 from .insert import Insert
@@ -13,4 +15,5 @@ from .transaction import Transaction
 from .update import Update
 
 __all__ = ("Create", "Delete", "Insert", "Live", "Remove", "Select", "Show", "Update", "Transaction", "DefineParam",
-           "DefineUser", "DefineEvent", "RebuildIndex", "DefineAnalyzer")
+           "DefineUser", "DefineEvent", "RebuildIndex", "DefineAnalyzer", "DefineAccessBearer", "DefineAccessJwt",
+           "DefineAccessRecord", "DefineField", "DefineScope", "DefineTable", "DefineToken", "DefineIndex")