+++ tests/integration_tests/test_ws_connection.py
@@ -436,10 +436,10 @@ class TestWebSocketConnection(TestCase):
         with surreal.connect() as connection:
             res = connection.root_info()
             self.assertFalse(res.is_error(), res)
-            self.assertEqual(len(res.result), 4)
+            self.assertEqual(len(res.result), 5)
             res = connection.root_info(structured=True)
             self.assertFalse(res.is_error(), res)
-            self.assertEqual(len(res.result), 4)
+            self.assertEqual(len(res.result), 5)
 
     def test_info_ns(self):
         surreal = Surreal(URL, 'test', credentials=('user_ns', 'user_ns'))