+++ tests/unit_tests/test_database.py
@@ -3,8 +3,9 @@ from unittest import TestCase, main
 from surrealist import Where, Algorithm
 from surrealist.ql.statements import Create, Update, Select
 from surrealist.ql.statements.define import DefineEvent, DefineParam, DefineScope, \
-    DefineIndex, DefineToken, DefineTable, DefineField
+    DefineIndex, DefineToken, DefineTable
 from surrealist.ql.statements.define_analyzer import DefineAnalyzer
+from surrealist.ql.statements.define_field import DefineField
 from surrealist.ql.statements.define_user import DefineUser
 from surrealist.ql.statements.define_access import DefineAccessJwt, DefineAccessRecord
 from surrealist.ql.statements.transaction import Transaction